+++ proofs/src/plonk/lookup/verifier.rs
@@ -5,7 +5,7 @@ use ff::{PrimeField, WithSmallOrderMulGroup};
 use super::{super::circuit::Expression, Argument};
 use crate::{
     plonk::{Error, VerifyingKey},
-    poly::{commitment::PolynomialCommitmentScheme, Rotation, VerifierQuery},
+    poly::{commitment::PolynomialCommitmentScheme, CommitmentLabel, Rotation, VerifierQuery},
     transcript::{Hashable, Transcript},
 };
 
@@ -186,30 +186,35 @@ impl<F: WithSmallOrderMulGroup<3>, CS: PolynomialCommitmentScheme<F>> Evaluated<
             // Open lookup product commitment at x
             .chain(Some(VerifierQuery::new(
                 x,
+                CommitmentLabel::NoLabel,
                 &self.committed.product_commitment,
                 self.product_eval,
             )))
             // Open lookup input commitments at x
             .chain(Some(VerifierQuery::new(
                 x,
+                CommitmentLabel::NoLabel,
                 &self.committed.permuted.permuted_input_commitment,
                 self.permuted_input_eval,
             )))
             // Open lookup table commitments at x
             .chain(Some(VerifierQuery::new(
                 x,
+                CommitmentLabel::NoLabel,
                 &self.committed.permuted.permuted_table_commitment,
                 self.permuted_table_eval,
             )))
             // Open lookup input commitments at \omega^{-1} x
             .chain(Some(VerifierQuery::new(
                 x_inv,
+                CommitmentLabel::NoLabel,
                 &self.committed.permuted.permuted_input_commitment,
                 self.permuted_input_inv_eval,
             )))
             // Open lookup product commitment at \omega x
             .chain(Some(VerifierQuery::new(
                 x_next,
+                CommitmentLabel::NoLabel,
                 &self.committed.product_commitment,
                 self.product_next_eval,
             )))