+++ proofs/src/plonk/trash/verifier.rs
@@ -3,7 +3,7 @@ use ff::{PrimeField, WithSmallOrderMulGroup};
 use super::Argument;
 use crate::{
     plonk::{Error, Expression},
-    poly::{commitment::PolynomialCommitmentScheme, VerifierQuery},
+    poly::{commitment::PolynomialCommitmentScheme, CommitmentLabel, VerifierQuery},
     transcript::{Hashable, Transcript},
 };
 
@@ -83,6 +83,7 @@ impl<F: WithSmallOrderMulGroup<3>, CS: PolynomialCommitmentScheme<F>> Evaluated<
     pub(crate) fn queries(&self, x: F) -> impl Iterator<Item = VerifierQuery<F, CS>> + Clone {
         vec![VerifierQuery::new(
             x,
+            CommitmentLabel::NoLabel,
             &self.committed.trash_commitment,
             self.trash_eval,
         )]