+++ proofs/src/plonk/verifier.rs
@@ -5,7 +5,7 @@ use ff::{FromUniformBytes, WithSmallOrderMulGroup};
 use super::{vanishing, Error, VerifyingKey};
 use crate::{
     plonk::traces::VerifierTrace,
-    poly::{commitment::PolynomialCommitmentScheme, VerifierQuery},
+    poly::{commitment::PolynomialCommitmentScheme, CommitmentLabel, VerifierQuery},
     transcript::{read_n, Hashable, Sampleable, Transcript},
     utils::arithmetic::compute_inner_product,
 };
@@ -423,6 +423,7 @@ where
                             if column.index() < nb_committed_instances {
                                 Some(VerifierQuery::new(
                                     vk.domain.rotate_omega(x, at),
+                                    CommitmentLabel::Instance(column.index()),
                                     &committed_instances[column.index()],
                                     instance_evals[query_index],
                                 ))
@@ -435,6 +436,7 @@ where
                         move |(query_index, &(column, at))| {
                             VerifierQuery::new(
                                 vk.domain.rotate_omega(x, at),
+                                CommitmentLabel::Advice(column.index()),
                                 &advice_commitments[column.index()],
                                 advice_evals[query_index],
                             )
@@ -449,6 +451,7 @@ where
             vk.cs.fixed_queries.iter().enumerate().map(|(query_index, &(column, at))| {
                 VerifierQuery::new(
                     vk.domain.rotate_omega(x, at),
+                    CommitmentLabel::Fixed(column.index()),
                     &vk.fixed_commitments[column.index()],
                     fixed_evals[query_index],
                 )