+++ proofs/src/poly/kzg/mod.rs
@@ -34,7 +34,7 @@ use crate::{
             params::{ParamsKZG, ParamsVerifierKZG},
             utils::construct_intermediate_sets,
         },
-        query::VerifierQuery,
+        query::{CommitmentLabel, VerifierQuery},
         Coeff, Error, LagrangeCoeff, Polynomial, ProverQuery,
     },
     transcript::{Hashable, Sampleable, Transcript},
@@ -212,7 +212,7 @@ where
                 None
             };
             for (scalar, commitment) in com_data.commitment.as_terms(eval_point_opt) {
-                msm.append_term(scalar, commitment);
+                msm.append_term(scalar, commitment, com_data.commitment_label.clone());
             }
             q_coms[com_data.set_index].push(msm);
             q_eval_sets[com_data.set_index].push(com_data.evals);
@@ -272,7 +272,7 @@ where
             let mut coms = q_coms;
             let mut f_com_as_msm = MSMKZG::init();
 
-            f_com_as_msm.append_term(E::Fr::ONE, f_com);
+            f_com_as_msm.append_term(E::Fr::ONE, f_com, CommitmentLabel::NoLabel);
             coms.push(f_com_as_msm);
 
             #[cfg(feature = "truncated-challenges")]
@@ -300,12 +300,16 @@ where
         let pi: E::G1 = transcript.read().map_err(|_| Error::SamplingError)?;
 
         let mut pi_msm = MSMKZG::<E>::init();
-        pi_msm.append_term(E::Fr::ONE, pi);
+        pi_msm.append_term(E::Fr::ONE, pi, CommitmentLabel::Custom("π".into()));
 
-        // Scale zπ -vG
+        // Scale zπ - vG
         let scaled_pi = MSMKZG {
             scalars: vec![x3, v],
             bases: vec![pi, -E::G1::generator()],
+            labels: vec![
+                CommitmentLabel::Custom("π".into()),
+                CommitmentLabel::Custom("-G".into()),
+            ],
         };
 
         // (π, C − vG + zπ)
@@ -336,7 +340,7 @@ mod tests {
                 KZGCommitmentScheme,
             },
             query::{ProverQuery, VerifierQuery},
-            EvaluationDomain,
+            CommitmentLabel, EvaluationDomain,
         },
         transcript::{CircuitTranscript, Hashable, Sampleable, Transcript},
         utils::arithmetic::eval_polynomial,
@@ -379,15 +383,16 @@ mod tests {
         let bvx: E::Fr = transcript.read().unwrap();
         let cvy: E::Fr = transcript.read().unwrap();
 
+        use CommitmentLabel::NoLabel;
         let valid_queries = std::iter::empty()
-            .chain(Some(VerifierQuery::new(x, &a, avx)))
-            .chain(Some(VerifierQuery::new(x, &b, bvx)))
-            .chain(Some(VerifierQuery::new(y, &c, cvy)));
+            .chain(Some(VerifierQuery::new(x, NoLabel, &a, avx)))
+            .chain(Some(VerifierQuery::new(x, NoLabel, &b, bvx)))
+            .chain(Some(VerifierQuery::new(y, NoLabel, &c, cvy)));
 
         let invalid_queries = std::iter::empty()
-            .chain(Some(VerifierQuery::new(x, &a, avx)))
-            .chain(Some(VerifierQuery::new(x, &b, avx)))
-            .chain(Some(VerifierQuery::new(y, &c, cvy)));
+            .chain(Some(VerifierQuery::new(x, NoLabel, &a, avx)))
+            .chain(Some(VerifierQuery::new(x, NoLabel, &b, avx)))
+            .chain(Some(VerifierQuery::new(y, NoLabel, &c, cvy)));
 
         let queries = if should_fail {
             invalid_queries